Assigned devices

+++ app/Http/Controllers/AssignedDevicesController.php
@@ -17,14 +17,12 @@ class AssignedDevicesController extends Controller
     public function index()
     {   
         //Get Free Devices For Assign
-        $devices = Devices::where('status',1)->get();
-        // dd($devices);
+        $devices = Devices::where('status',0)->get();
         //Get Users Without Admin To Assign Device Where Users Are Active 
         $users = Users::with('department')->whereHas('role', function($q){
             $q->where('name', '!=', 'Super Admin');
         })->where('status',1)->get();
         $assignedDevices = AssignedDevices::with('user','device')->get();
-        // dd($assignedDevices);
 
         return view('devices.assigned.index', compact('devices','users','assignedDevices'));
     }
@@ -43,31 +41,34 @@ class AssignedDevicesController extends Controller
      */
     public function store(Request $request)
     {
-        // dd($request);
         $validator = \Validator::make($request->all(), [
             'device_id' => 'required', 
             'user_id' => 'required',
             'assigned_from' => 'required',
-            'assigned_to' => 'nullable',  
+            'assigned_to' => 'nullable|after_or_equal:assigned_from|before:today',  
         ]);        
         if ($validator->fails())
         {
             return response()->json(['errors'=>$validator->errors()->all()]);
         }
         $validate = $validator->valid();
-        
+        $status = 1;
+        if ($request->assigned_to != null &&  $request->assigned_to == $request->assigned_from) {
+            $status = 0;
+        }
         $deviceassigned = AssignedDevices::Create([
             'device_id' => $validate['device_id'],
             'user_id' => $validate['user_id'],
             'from' => $validate['assigned_from'],
             'to' => $validate['assigned_to'],
+            'status' => $status,
             'created_at' => date('Y-m-d H:i:s'),
             'updated_at' => date('Y-m-d H:i:s'),
         ]);
         if($deviceassigned){
             Devices::where('id', $validate['device_id'])->update(
                 [
-                    'status' => 0,
+                    'status' => $status,
                 ]);
         }
 
@@ -102,7 +103,7 @@ class AssignedDevicesController extends Controller
         $inUseDevices = [];
 
         foreach ($devices as $device) {
-            if ($device->status == 1) {
+            if ($device->status == 0) {
                 $freeDevices[] = $device;
             } else {
                 $inUseDevices[] = $device;
@@ -122,10 +123,41 @@ class AssignedDevicesController extends Controller
      * @param  \App\Models\AssignedDevices  $assignedDevices
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, AssignedDevices $assignedDevices)
+    public function update(Request $request,$id)
     {
-        //
+        $validator = \Validator::make($request->all(), [
+            'edit_device_id' => 'required',   
+            'edit_user_id' => 'required',   
+            'edit_assigned_from' => 'required',   
+            'edit_assigned_to' => 'nullable|after_or_equal:edit_assigned_from|before:today',      
+        ]);
+ 
+        if ($validator->fails())
+        {
+            return response()->json(['errors'=>$validator->errors()->all()]);
+        }
+        $status = 1;
+		if ($request->edit_assigned_to != null) {
+            $status = 0;
+        }
+      $assignedDevice =  AssignedDevices::where('id', $id)
+        ->update([
+            'device_id' => $request->edit_device_id,
+            'user_id' => $request->edit_user_id,
+            'from' => $request->edit_assigned_from,
+            'to' => $request->edit_assigned_to,
+            'status' => $status,
+        ]);
+        if($assignedDevice){
+            $deviceStatus = Devices::where('id', $request->edit_device_id)->update([
+                'status' => $status,
+            ]);
+        }
+		
+        $request->session()->flash('message','Assigned Device updated successfully.');
+        return redirect()->route('devices.assigned.index')->with('assignedDevices', $assignedDevice);
     }
+
 
     /**
      * Remove the specified resource from storage.
@@ -136,15 +168,20 @@ class AssignedDevicesController extends Controller
     public function destroy(Request $request)
     {   
         $device = AssignedDevices::where('id',$request->id)->first();
-        $device_id = $device->device_id;
-        $device->delete();
-        if ($device->deleted_at != null) {
-            Devices::where('id', $device_id)->update(
-                [
-                    'status' => 1,
-                ]);
+        if($device->to == null){
+        $request->session()->flash('error','You Cannot Delete Assigned Device without Adding To Date.');
+        }else{
+            $device_id = $device->device_id;
+            $device->status = 0;
+            $device->delete();
+            if ($device->deleted_at != null) {
+                Devices::where('id', $device_id)->update(
+                    [
+                        'status' => 0,
+                    ]);
+            }
+            $request->session()->flash('message','Assigned Device deleted successfully.');
         }
-        $request->session()->flash('message','Assigned Device deleted successfully.');
        return Response()->json($device);
     }
 }
